pyproject.toml
[project]
name = "aws-capture-auto"
version = "0.1.0"
description = "AWSへのログイン・スクリーンショット取得・リソース取得を自動化するツール"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "playwright>=1.44.0",
    "pytest-playwright>=0.4.4",
    "boto3>=1.34.63",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
aws-capture = "main:main"
